completion of cancel order and modify order functions and added simple tests in main for adding a new order; also added print functions

+++ include/Using.h
@@ -3,6 +3,6 @@
 #include <cstdint>
 
 using OrderId = uint64_t;
-using Price = uint64_t;
+using Price = double;
 using Quantity = uint64_t;
 using Timestamp = uint64_t;
